move bamstats to the generate tool
